+++ src/Transferable.js
@@ -6,6 +6,7 @@ function Transferable (obj, transferables) {
   this.obj = obj
   const transfers = transferables || obj
   this.transferables = [].concat(transfers).map((value) => {
+    // TODO: Add Float64Array
     if (value instanceof Uint8Array || value instanceof Uint16Array || value instanceof Uint32Array) {
       return value.buffer
     }