+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 ## Version 0.2.0
-- Puddle internal thread id is exposed to worker via `const { threadId } = require('thread-puddle')`
+- Puddle internal thread id is exposed to worker via `const { threadId } = require('thread-puddle')`.
+- The module now exports `createThreadPool` as a pool constructor.